TW-1941: Separate chat list slidable actions
